backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
pydantic-settings